+++ Riot/Modules/Room/RoomInfo/RoomInfoList/RoomInfoListCoordinatorType.swift
@@ -22,6 +22,7 @@ protocol RoomInfoListCoordinatorDelegate: AnyObject {
     func roomInfoListCoordinator(_ coordinator: RoomInfoListCoordinatorType, wantsToNavigateTo target: RoomInfoListTarget)
     func roomInfoListCoordinatorDidCancel(_ coordinator: RoomInfoListCoordinatorType)
     func roomInfoListCoordinatorDidLeaveRoom(_ coordinator: RoomInfoListCoordinatorType)
+    func roomInfoListCoordinatorDidRequestReportRoom(_ coordinator: RoomInfoListCoordinatorType)
 }
 
 /// `RoomInfoListCoordinatorType` is a protocol describing a Coordinator that handle key backup setup passphrase navigation flow.