+++ Riot/Utils/Tools.h
@@ -48,4 +48,11 @@
  */
 + (NSURL*)fixURLWithSeveralHashKeys:(NSURL*)url;
 
+#pragma mark - Time utilities
+
+/**
+ * Convert a number of days to a duration in ms.
+ */
++ (uint64_t)durationInMsFromDays:(uint)days;
+
 @end